+++ src/dm-player/interfaces/infPlayer.ts
@@ -23,20 +23,11 @@ export default interface InfPlayer {
     //xid of videos to exclude separated by ","
     excludeIds: string;
 
-    //search in a playlit ? provide playlist xid OR false
+    //search in a playlist ? provide playlist xid OR false
     searchInPlaylist: string | boolean;
 
     //syndication key
     syndication: string;
-
-    //video auto play
-    autoPlayMute: boolean;
-
-    //enable video queue
-    queueEnable: boolean;
-
-    //enable video queue autoplay next
-    queueEnableNext: boolean;
 
     //shows player controls
     controls: boolean;
@@ -47,7 +38,54 @@ export default interface InfPlayer {
     //You can get CPE ID from partner HQ on template tabs, inside embed menu
     cpeId: string[];
 
+    // Add selector to get keywords meta
     keywordsSelector: string;
+
+    // Assign false the video inventory is not up to date
+    getUpdatedVideo?: boolean;
+
+    // Title before video player. I.E. "see also:"
+    preVideoTitle?: string;
+
+    // Show video title or not, set true if need to show video title
+    showVideoTitle?: boolean;
+
+    // Info of the video in a card below the video player
+    showInfoCard?: boolean;
+
+    ////// Config to player
+
+    // Picture-in-Picture at start
+    pipAtStart?: boolean;
+
+    // No scroll to play
+    noStp?: boolean;
+
+    // No Picture-in-Picture
+    noPip?: boolean;
+
+    //enable video queue
+    queueEnable?: boolean;
+
+    //enable video queue autoplay next
+    queueEnableNext?: boolean;
+
+    ///// Config to the CPE script
+
+    // Set video to pause when it is not in viewport
+    scrollToPause?: boolean;
+
+    // Activate sound
+    stpSound?: boolean;
+
+    // Video auto play
+    autoPlayMute: boolean;
+
+    // Custom player style
+    playerStyleEnable: boolean;
+
+    // Define player style
+    playerStyleColor: string;
 
     width?: number;
     height?: number;