+++ src/dm-player/interfaces/infVideo.ts
@@ -1,4 +1,6 @@
 export default interface InfVideo {
     id: string;
     title?: string;
+    description?: string;
+    'owner.avatar_190_url'?: string;
 }