+++ src/entries/dm-embed.ts
@@ -1,6 +1,12 @@
 import DmPlayer from '../dm-player/dm-player';
+import { waitFor } from '../utilities/wait-for';
 
-const el: NodeListOf<HTMLDivElement> = document.querySelectorAll('.dm-player');
+const init = async () => {
 
-const dmEmbed = new DmPlayer(el);
+	await waitFor( () => document.querySelectorAll('.dm-player').length > 0, 500, 2000, "Timeout to get DM placeholder");
+	const el: NodeListOf<HTMLDivElement> = document.querySelectorAll('.dm-player');
 
+	const dmEmbed = new DmPlayer(el);
+};
+
+init();