+++ CHANGELOG.md
@@ -1,5 +1,47 @@
 # Changelog
 
+## [1.1.6] (2020-04-27)
+
+**Fixed Bugs:**
+- Fixed publisher logo on info card
+
+**Implemented enhancements:**
+- Remove @types/dotenv package
+- Add version badge and license badge to readme
+- Add new param to handle updated video
+
+## [1.1.5] (2020-04-24)
+
+**Fixed Bugs:**
+- Add waiting function to wait for the element is ready or not
+
+**Implemented enhancements:**
+- Search API improvements
+
+## [1.1.2] (2020-04-22)
+
+**Fixed Bugs:**
+- Fixed get recent video
+
+**Implemented enhancements:**
+- Add owner avatar to info card
+
+## [1.1.1] (2020-04-22)
+
+**Fixed Bugs:**
+- Handle empty language params
+
+**Implemented enhancements:**
+- Add info card below the video
+- Add animation while showing the video
+
+## [1.1.0] (2020-04-16)
+
+**Implemented enhancements:**
+- Handle many parameters in tag placeholder
+- Add before title i.e. "See also:"
+- Add video title below the video
+
 ## [1.0.0] (2020-03-23)
 
 **Features:**